+++ packages/frameworks/test/frameworks.unit.test.ts
@@ -218,6 +218,7 @@ describe('frameworks', () => {
   const skipExamples = [
     'dojo',
     'fastify', // FIXME: fastify once builder lands in builder container because integration tests will fail if there's an example with no tests
+    'elysia', // FIXME: elysia once builder lands in builder container because integration tests will fail if there's an example with no tests
     'saber',
     'gridsome',
     'sanity-v3',